requirements.txt
streamlit==1.25.0
langchain==0.0.270
openai~=0.27.8
python-dotenv~=1.0.0
tiktoken~=0.4.0
pandas~=2.0.3
tenacity==8.2.3
pydantic==2.2.1
aiohttp==3.8.5
aioprocessing~=2.0.1
numpy==1.25.2
python-dateutil~=2.8.2
stanza~=1.5.0
